Bills Done

+++ demo/demo/src/main/java/com/sevitours/demo/bill/service/UpdateBillCommand.java
@@ -1,13 +1,14 @@
 package com.sevitours.demo.bill.service;
+
 import com.sevitours.demo.bill.model.Bill;
 import lombok.Getter;
 
 @Getter
 public class UpdateBillCommand {
-    private Integer id;
+    private Long id;
     private Bill bill;
 
-    public UpdateBillCommand(Bill bill, Integer id) {
+    public UpdateBillCommand(Bill bill, Long id) {
         this.bill = bill;
         this.id = id;
     }
